test: added more test cases to improve coverage and meet SonarQube quality gate
